ryv_rentas/requirements.txt
Django==4.2.11
reportlab==4.1.0
whitenoise==6.6.0
python-decouple==3.8
gunicorn==21.2.0
